fix(cli): reorganize task cli module structure
Move task CLI functions from task_cli.py to task_cli_functions.py
Update imports and entry point in pyproject.toml
Remove old main.py and use new taskcli.py as entry point
