+++ pyproject.toml
@@ -9,4 +9,4 @@ dependencies = [
 ]
 
 [project.scripts]
-taskcli = "main:main"
+taskcli = "taskcli:main"
